Refactor analyse components ws and chart data
