+++ frontend/src/components/analyse/GraphOptions.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react';
+import React, { useState } from 'react';
 import styled from 'styled-components';
 import { ButtonPrimary } from '../styled';
 import { LuChartLine } from 'react-icons/lu';
@@ -10,8 +10,7 @@ import { useSelector } from 'react-redux';
 import { RootState } from '../../redux/store';
 import { GetUserFavoriteCryptosForAnalyseQuery } from '../../graphql/generated';
 import AiAnalyseModal from './AiAnalyseModal';
-import useWebSocketClient from '../../ws/useWebSocketClient';
-import { StompSubscription } from '@stomp/stompjs';
+import useWebSocket from '../../ws/useWebSocket';
 
 
 const OptionsContainer = styled.div`
@@ -103,26 +102,9 @@ const GraphOptions: React.FC<GraphOptionsProps> = ({ onOptionsChange, options })
     const { user } = useSelector((state: RootState) => state.auth);
     const [isAiAnalyseModalOpen, setIsAiAnalyseModalOpen] = useState(false);
     const [startAiAnalyse] = useMutation(START_AI_ANALYSE);
-    const [messages, setMessages] = useState<string[]>([]);
-    const { subscribe, unsubscribe, sendMessage, connected } = useWebSocketClient(import.meta.env.VITE_WS_API_URL); 
-
-    useEffect(() => {
-        let subscription: StompSubscription | null = null;
-        if (connected) {
-            subscription = subscribe(`/analyse/${user?.id}`, (message: string) => {
-                setMessages(prevMessages => [...prevMessages, message]);
-            }) as StompSubscription;
-        }
-
-        return () => {
-            if (subscription) { 
-                unsubscribe(subscription);
-            }
-        };
-    }, [connected, subscribe, unsubscribe]);
+    const { messages } = useWebSocket(`/analyse/${user?.id}`);
 
     const handleAiAnalyseModal = async () => {
-        console.log("startAiAnalyse", optionsState.cryptocurrency, user?.id);
         setIsAiAnalyseModalOpen(!isAiAnalyseModalOpen);
         await startAiAnalyse({
             variables: {
@@ -241,7 +223,6 @@ const GraphOptions: React.FC<GraphOptionsProps> = ({ onOptionsChange, options })
             <AiAnalyseModal
                 isOpen={isAiAnalyseModalOpen}
                 onClose={handleAiAnalyseModal}
-                cryptoId={optionsState.cryptocurrency}
                 messages={messages}
             />
         </OptionsContainer>
